Share location between compiler crates
