+++ compiler/core/src/lib.rs
@@ -1,8 +1,10 @@
 #![doc(html_logo_url = "https://raw.githubusercontent.com/RustPython/RustPython/main/logo.png")]
 #![doc(html_root_url = "https://docs.rs/rustpython-compiler-core/")]
 
-pub mod bytecode;
+mod bytecode;
+mod location;
 mod mode;
 
 pub use bytecode::*;
+pub use location::Location;
 pub use mode::Mode;
